requirements.txt
appdirs==1.4.0
Django==1.10.5
django-crispy-forms==1.6.1
django-filter==1.0.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
Markdown==2.6.8
markdown2==2.3.2
packaging==16.8
PyJWT==1.4.2
pyparsing==2.1.10
six==1.10.0
